Added most Act 3 bosses and Fist Sinner.
Added the 4 memory bosses (The singer, clover dancers, crust king and
Nyleth). Seth, Crow Court and hunter Gurr (Gurr normally requires his
quest to be fought but since the area unlocks permenantly when you find
it in the quest, and the item for turning in the quest is in the room
next to Gurr, it's possible to respawn the fight normally).

Also added First Sinner, who requires removal of Rune Rage. The game
doesn't handle removing silk skills well, and even though the skill
isn't always removed, you can always fight First Sinner after the
"hasSilkBomb" is set to false.

Also added the Unraveled and Forebrothers bosses.